check zero buffer update, reformat pipeline layout
